Integrate pipeline packages
