+++ pipeline/__init__.py
@@ -1,11 +1,40 @@
-from .base_pipeline import BasePruningPipeline
-from .pruning_pipeline import PruningPipeline
-from pruning_pipeline.context import PipelineContext
-from pruning_pipeline.step import PipelineStep
+from importlib import import_module
 
 __all__ = [
     "BasePruningPipeline",
     "PruningPipeline",
     "PipelineContext",
     "PipelineStep",
+    "LoadModelStep",
+    "TrainStep",
+    "AnalyzeModelStep",
+    "GenerateMasksStep",
+    "ApplyPruningStep",
+    "ReconfigureModelStep",
+    "CalcStatsStep",
+    "CompareModelsStep",
 ]
+
+
+def __getattr__(name: str):
+    if name == "BasePruningPipeline":
+        return import_module("pipeline.base_pipeline").BasePruningPipeline
+    if name == "PruningPipeline":
+        return import_module("pipeline.pruning_pipeline").PruningPipeline
+    if name == "PipelineContext":
+        return import_module("pipeline.context").PipelineContext
+    if name == "PipelineStep":
+        return import_module("pipeline.step").PipelineStep
+    if name in {
+        "LoadModelStep",
+        "TrainStep",
+        "AnalyzeModelStep",
+        "GenerateMasksStep",
+        "ApplyPruningStep",
+        "ReconfigureModelStep",
+        "CalcStatsStep",
+        "CompareModelsStep",
+    }:
+        module = import_module(f"pipeline.step.{name.lower()}")
+        return getattr(module, name)
+    raise AttributeError(name)
