+++ pipeline/step/generate_masks.py
